move lsst_camera() into its own module
so that Zernike correction to pupil->focal transformation
can import from DMtoCameraModule

move DMtoCameraModule import up
