+++ python/lsst/sims/coordUtils/__init__.py
@@ -1,3 +1,4 @@
+from .LsstCameraMethod import *
+from .DMtoCameraModule import *
 from .CameraUtils import *
 from .LsstCameraUtils import *
-from .DMtoCameraModule import *
